+++ azimuth/utils/filtering.py
@@ -14,6 +14,7 @@ def filter_dataset_split(
     dataset_split: Dataset,
     filters: Union[DatasetFilters, NamedDatasetFilters],
     config: ProjectConfig,
+    without_postprocessing: bool = False,
 ) -> Dataset:
     """Filter dataset_split according to a filter component.
 
@@ -21,6 +22,7 @@ def filter_dataset_split(
         dataset_split: examples to filter.
         filters: On what to filter on.
         config: Azimuth Config.
+        without_postprocessing: Filter on columns without_postprocessing (model)
 
     Returns:
         Filtered dataset_split.
@@ -29,12 +31,14 @@ def filter_dataset_split(
         ValueError, if required column to filter by is not in the dataset_split's columns.
 
     """
-
     if filters.confidence_min > 0 or filters.confidence_max < 1:
+        confidence_column = (
+            DatasetColumn.model_confidences
+            if without_postprocessing
+            else DatasetColumn.postprocessed_confidences
+        )
         dataset_split = dataset_split.filter(
-            lambda x: filters.confidence_min
-            <= x[DatasetColumn.postprocessed_confidences][0]
-            <= filters.confidence_max
+            lambda x: filters.confidence_min <= x[confidence_column][0] <= filters.confidence_max
         )
     if len(filters.labels) > 0:
         if config.columns.label not in dataset_split.column_names:
@@ -46,13 +50,21 @@ def filter_dataset_split(
         by = filters.utterance.lower()
         dataset_split = dataset_split.filter(lambda x: by in x[config.columns.text_input].lower())
     if len(filters.predictions) > 0:
-        if DatasetColumn.postprocessed_prediction not in dataset_split.column_names:
-            raise ValueError(
-                f"{DatasetColumn.postprocessed_prediction} not in dataset_split's column names"
-            )
-        dataset_split = dataset_split.filter(
-            lambda x: x[DatasetColumn.postprocessed_prediction] in filters.predictions
+        prediction_column = (
+            DatasetColumn.model_predictions
+            if without_postprocessing
+            else DatasetColumn.postprocessed_prediction
         )
+        if prediction_column not in dataset_split.column_names:
+            raise ValueError(f"{prediction_column} not in dataset_split's column names")
+        if without_postprocessing:
+            dataset_split = dataset_split.filter(
+                lambda x: x[DatasetColumn.model_predictions][0] in filters.predictions
+            )
+        else:
+            dataset_split = dataset_split.filter(
+                lambda x: x[DatasetColumn.postprocessed_prediction] in filters.predictions
+            )
     if len(filters.data_actions) > 0:
         # We do OR for data_action tags.
         dataset_split = dataset_split.filter(
@@ -62,8 +74,13 @@ def filter_dataset_split(
             )
         )
     if len(filters.outcomes) > 0:
+        outcome_column = (
+            DatasetColumn.model_outcome
+            if without_postprocessing
+            else DatasetColumn.postprocessed_outcome
+        )
         # We do OR for outcomes.
-        dataset_split = dataset_split.filter(lambda x: x[DatasetColumn.outcome] in filters.outcomes)
+        dataset_split = dataset_split.filter(lambda x: x[outcome_column] in filters.outcomes)
     if len(filters.smart_tags) > 0:
         # We do AND for smart tags.
         dataset_split = dataset_split.filter(