+++ tests/utils.py
@@ -2,6 +2,7 @@
 # This source code is licensed under the Apache 2.0 license found in the LICENSE file
 # in the root directory of this source tree.
 from azimuth.modules.model_contract_task_mapping import model_contract_task_mapping
+from azimuth.modules.model_performance.outcomes import OutcomesModule
 from azimuth.types import DatasetSplitName, ModuleOptions, SupportedMethod
 
 
@@ -16,3 +17,14 @@ def save_predictions(config, ds_split_name=DatasetSplitName.eval):
     pred_res = pred_mod.compute_on_dataset_split()
     dm = pred_mod.get_dataset_split_manager()
     pred_mod.save_result(pred_res, dm)
+
+
+def save_outcomes(config, ds_split_name=DatasetSplitName.eval):
+    outcome_mod = OutcomesModule(
+        dataset_split_name=ds_split_name,
+        config=config,
+        mod_options=ModuleOptions(pipeline_index=0),
+    )
+    outcome_res = outcome_mod.compute_on_dataset_split()
+    dm = outcome_mod.get_dataset_split_manager()
+    outcome_mod.save_result(outcome_res, dm)